service/requirements.txt
boto3==1.26.40
botocore==1.29.40
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
protobuf==4.21.2
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
slack-sdk==3.19.5
urllib3==1.26.13
Werkzeug==2.2.2
